requirements.txt
mcp[cli]
azure-identity
azure-monitor-query
azure-mgmt-monitor
